wip: backmerge prep release and some cleaning
